Clean up some code
